SupportArmCommand solved
